+++ api/src/api/translation.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 const { errorWrap } = require('../middleware');
 const Translation = require('../models/translation');
+const VerifiedTranslation = require('../models/verifiedTranslation');
 
 // Get translations of specified language
 router.get(
@@ -25,6 +26,70 @@ router.get(
         result: null,
       });
     }
+  }),
+);
+
+// Report a translation error for an ID
+router.patch(
+  '/report/:id',
+  errorWrap(async (req, res) => {
+    const { id } = req.params;
+    const { language, type } = req.body;
+
+    const updateQuery = { $inc: { numReports: 1 } };
+
+    switch (type) {
+      case 'resource':
+        await VerifiedTranslation.updateOne(
+          {
+            resourceID: id,
+            language,
+          },
+          updateQuery,
+        );
+        break;
+      case 'category':
+        await VerifiedTranslation.updateOne(
+          {
+            categoryID: id,
+            language,
+          },
+          updateQuery,
+        );
+        break;
+      case 'subcategory':
+        await VerifiedTranslation.updateOne(
+          {
+            subcategoryID: id,
+            language,
+          },
+          updateQuery,
+        );
+        break;
+      case 'testimonial':
+        await VerifiedTranslation.updateOne(
+          {
+            testimonialID: id,
+            language,
+          },
+          updateQuery,
+        );
+        break;
+      default:
+        res.json({
+          code: 400,
+          message: 'Could not find specified type',
+          success: false,
+          result: null,
+        });
+    }
+
+    res.json({
+      code: 200,
+      message: 'Successfully reported an error',
+      success: true,
+      result: null,
+    });
   }),
 );
